refactor: menu components folder structure
